+++ jdaviz/configs/default/plugins/markers/markers.py
@@ -208,6 +208,7 @@ class Markers(PluginTemplateMixin, ViewerSelectMixin, TableMixin):
             else:
                 viewer.remove_event_callback(callback)
 
+    # this is where items are being added to the table
     def _on_viewer_key_event(self, viewer, data):
         if data['event'] == 'keydown' and data['key'] == 'm':
             row_info = self.coords_info.as_dict()